# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a highly scalable URL shortener API built with Go, Redis, and Cassandra.",
        "title": "Shorner URL API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/shorten": {
            "post": {
                "description": "Creates a short URL for the provided original URL. If it already exists, returns the existing code.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Shorten a URL",
                "parameters": [
                    {
                        "description": "URL to shorten",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/shorner.shortenRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/shorner.ShortenOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/shorner.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/shorner.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/urls/{code}": {
            "get": {
                "description": "Returns information about a specific shortened URL.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Get URL Details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short URL Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shorner.URL"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/shorner.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/shorner.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/shorner.errorResponse"
                        }
                    }
                }
            }
        },
        "/{code}": {
            "get": {
                "description": "Redirects the client to the original URL associated with the short code.",
                "tags": [
                    "urls"
                ],
                "summary": "Redirect to Original URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short URL Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "301": {
                        "description": "Moved Permanently"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/shorner.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/shorner.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/shorner.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "shorner.ShortenOutput": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "original_url": {
                    "type": "string"
                },
                "short_url": {
                    "type": "string"
                }
            }
        },
        "shorner.URL": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "original_url": {
                    "type": "string"
                }
            }
        },
        "shorner.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid url: must be a valid http or https url"
                }
            }
        },
        "shorner.shortenRequest": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string",
                    "example": "https://github.com/joaovv-Vitor/Shorner-url"
                }
            }
        }
    }
}